First commit
Currently able to display tetriminos to the screen in what should be a reasonable area.  Currently implementing tetrimino rotation, it's basically done but needs checking.  The code base is significantly large enough and complicated that I feel it's time to start using source control.  The project could use some tidying and commenting but it's getting late...