Beginnings of game loop in place
Currently handling player input (but need to distinguish between pressing and holding to prevent rapid input); bare bones of game loop in place (i.e. a piece gets spawned, it drops slowly overtime and it becomes part of the underlying grid on downward collisions).  This needs some serious refactoring but the logic is in place for now.